# specs/001-mvp-travel-planner/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mimi Journey API
  description: 旅行規劃與軌跡紀錄 API
  version: 1.0.0
  contact:
    name: Mimi Journey Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mimi-journey.app/v1
    description: Production
...
paths:
...
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Google OAuth 回調
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: 登入成功，重定向至前端
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      summary: 取得當前用戶資訊
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 用戶資訊
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [auth]
      summary: 登出
      security:
        - cookieAuth: []
      responses:
        "204":
          description: 登出成功

  # ============ Calendar ============
...
  /places/search:
    get:
      tags: [places]
      summary: 搜尋地點
      security:
        - cookieAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: 搜尋關鍵字
        - name: location
          in: query
          schema:
            type: string
          description: "中心點座標 (lat,lng)"
          example: "25.0330,121.5654"
      responses:
        "200":
          description: 搜尋結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  places:
                    type: array
                    items:
                      $ref: "#/components/schemas/Place"
  /places/geocode:
    get:
      tags: [places]
      summary: 地址轉座標
      security:
        - cookieAuth: []
      parameters:
        - name: address
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 地理編碼結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeocodeResult"
  /places/reverse-geocode:
    get:
      tags: [places]
      summary: 座標轉地址
      security:
        - cookieAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 反向地理編碼結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
...
  responses:
    Unauthorized:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double

    # ---- Auth ----
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        picture_url:
          type: string
          format: uri

    # ---- Calendar ----
...
    Place:
      type: object
      properties:
        place_id:
          type: string
        name:
          type: string
        address:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        types:
          type: array
          items:
            type: string
    GeocodeResult:
      type: object
      properties:
        location:
          $ref: "#/components/schemas/Location"
        formatted_address:
          type: string
        place_id:
          type: string